Tabbing, select all in cells, and change single click to ctrl click
Thanks for this great npm for quill version 1. This module seems to work better than most of the Quill 2.0 table modules even, especially with the multi-line in cell!

I have added some minor optional key bindings which seem to fit nicely with what you have currently in keyboardHandler. I have added a tabbing bind so the user can quickly enter data move to the next cell or previous cell.

I have also added a custom optional select all for the cell alone, for quick delete, or edit of text. This works nicely with your implementation of multi-line per cell which has been appreciated.

Lastly, I added a different approach for cell select (previous single click). I appreciated that it was a separate function for click and double click but found it much easier as a ctrl click for selecting the cell. Regular click as the entry to cell to begin typing, etc.